refactor: #1 🛠️ divide workspace
