+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 dist
 .cspellcache
+bin
+temp
