+++ packages/backend/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@tegaki/backend",
+  "version": "0.0.1",
+  "description": "@tegaki/backend",
+  "main": "dist/index.js",
+  "keywords": [],
+  "author": "Ubugeeei <ubuge1122@gmail.com>",
+  "license": "MIT"
+}
