+++ packages/cli/src/cmd/datagen/index.d.ts
@@ -1,0 +1,1 @@
+export declare const datagen: import("citty").CommandDef<import("citty").ArgsDef>;
