+++ packages/cli/src/cmd/jTegaki/index.d.ts
@@ -1,0 +1,1 @@
+export declare const openJTegaki: import("citty").CommandDef<import("citty").ArgsDef>;
