+++ packages/cli/src/cmd/unicodegen/index.d.ts
@@ -1,0 +1,14 @@
+export declare const unicodegen: import("citty").CommandDef<{
+    map: {
+        type: "boolean";
+        alias: string[];
+        description: string;
+        required: false;
+    };
+    separator: {
+        type: "string";
+        alias: string[];
+        description: string;
+        default: string;
+    };
+}>;
