--- packages/dataset/index.ts
@@ -1,1 +1,0 @@
-export { KANJI_DATA_SET } from "./kanji";
