+++ packages/dataset/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@tegaki/dataset",
+  "version": "0.0.1",
+  "description": "",
+  "main": "dist/index.js",
+  "keywords": [],
+  "author": "Ubugeeei <ubuge1122@gmail.com>",
+  "license": "MIT"
+}
