+++ packages/dataset/src/index.d.ts
@@ -1,0 +1,1 @@
+export { KANJI_DATA_SET } from "./kanji";
