+++ packages/dataset/src/kanji.d.ts
@@ -1,0 +1,2 @@
+import { type TegakiStroke } from "tegaki";
+export declare const KANJI_DATA_SET: Array<TegakiStroke>;
