--- packages/tegaki/index.ts
@@ -1,3 +1,0 @@
-export * from "@tegaki/dataset";
-export * from "@tegaki/frontend";
-export * from "@tegaki/types";
